Upload Arduino code for RFID door lock.
I know that most of this code isn't mine, but I am nowhere near experienced enough to figure all of the RFID stuff out by myself. But it did take a while to find the right way to combine code for the RFID, the infrared sensor, the buttons, etc. 

(DISCLAIMER : Although the name of the file suggests that it works with MiFare Ultralight, it only works once, then you have to reset it to be able to use it again. It works indefinitely with MiFare Classic cards, though. I do not know why it does this, and I can't figure it out because I don't have a way to connect a computer to it until I get a laptop. Just wanted to make all this known.)