Added the code for my sleeve project
I don't have an official name for this project yet, but when I'm finished I will be open to suggestions. Currently does work very well because I was trying to get the MPU6050 to send an interrupt when shaken, but I couldn't seem to get it to work. So, this is where I'm at now